fix(folder-editor): Relocate createSkillFormModelFromSkillResolved because it is not callable in skill.router.ts
